+++ webcomponents/lazy-img/CHANGELOG.md
@@ -1,3 +1,21 @@
+<a name="1.0.0-alpha.4"></a>
+# 1.0.0-alpha.4 (2019-05-24)
+
+### Fix
+
+* on attributes change, observe image again and fix new url
+
+<a name="1.0.0-alpha.3"></a>
+# 1.0.0-alpha.3 (2019-05-24)
+
+### Breaking
+
+* move to the org scoped package `@deckdeckgo/lazy-img`
+
+### Features
+
+* add a couple of CSS4 variables to style the image
+
 <a name="1.0.0-alpha.2"></a>
 # 1.0.0-alpha.2 (2019-05-18)